+++ src/vm.h
@@ -43,7 +43,8 @@ enum class CompileOption : unsigned short {
     OP(TRACE_EXIT, (1u << 0u), "traceonexit") \
     OP(MONITOR   , (1u << 1u), "monitor") \
     OP(NULLGLOB  , (1u << 2u), "nullglob") \
-    OP(DOTGLOB   , (1u << 3u), "dotglob")
+    OP(DOTGLOB   , (1u << 3u), "dotglob")     \
+    OP(FASTGLOB  , (1u << 4u), "fastglob")
 
 // set/unset via 'shctl' command
 enum class RuntimeOption : unsigned short {